requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
whitenoise==6.6.0

# Production server
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
